requirements.txt
python-telegram-bot[ext]
SQLAlchemy
pillow